moving the mocks file inside servicesmocks dir
