+++ internal/services/servicesmocks/rpc_service_mocks.go
@@ -1,7 +1,8 @@
-package services
+package servicesmocks
 
 import (
 	"github.com/stellar/wallet-backend/internal/entities"
+	"github.com/stellar/wallet-backend/internal/services"
 	"github.com/stretchr/testify/mock"
 )
 
@@ -9,7 +10,7 @@ type RPCServiceMock struct {
 	mock.Mock
 }
 
-var _ RPCService = (*RPCServiceMock)(nil)
+var _ services.RPCService = (*RPCServiceMock)(nil)
 
 func (r *RPCServiceMock) SendTransaction(transactionXdr string) (entities.RPCSendTransactionResult, error) {
 	args := r.Called(transactionXdr)
